% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filter Data according to condition evidence on evidence.nodes}
\usage{
filterData(data, st1, st2, remove.na, evidence.nodes, evidence,
  silent = TRUE)
}
\arguments{
\item{type}{data}
}
\description{
Filter Data according to condition evidence on evidence.nodes
}
\author{
M.N. Legasa
}
