% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNormalCorrelation.R
\name{findNormalCorrelation}
\alias{findNormalCorrelation}
\title{Find normal correlation}
\usage{
findNormalCorrelation(r, p01e1, p11e1, p01e2, p11e2, max.error = 0.001,
  max.iter = 25, accuracy = 1e+06, debug = FALSE)
}
\arguments{
\item{r}{correlation}
}
\description{
Find normal correlation
}
\author{
M.N. Legasa
}
