% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNormalCovarianceMatrix.R
\name{findNormalCovarianceMatrix}
\alias{findNormalCovarianceMatrix}
\title{Find Normal Covariance Matrix}
\usage{
findNormalCovarianceMatrix(correlations, transition.probabilities,
  max.error = 0.001, max.iter = 25, debug = FALSE,
  accuracy = 1e+06, parallelize = TRUE, cluster.type = "PSOCK",
  n.cores = NULL)
}
\description{
Find Normal Covariance Matrix
}
\author{
M.N. Legasa
}
