% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTransitionProbabilities.R
\name{findTransitionProbabilities}
\alias{findTransitionProbabilities}
\title{Find Transition Probabilities}
\usage{
findTransitionProbabilities(day0, day1)
}
\description{
Find Transition Probabilities
}
\author{
M.N. Legasa
}
