% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleData.R
\name{handleData}
\alias{handleData}
\title{Handle data from climate4R package or free format}
\usage{
handleData(data, dates = NULL, stations.id = NULL)
}
\description{
Handle data from climate4R package or free format
}
\author{
M.N. Legasa
}
