% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureMatrix.R
\name{measureMatrix}
\alias{measureMatrix}
\title{Measure Matrix}
\usage{
measureMatrix(data, measure = "phiCoef", mark.diagonal.na = TRUE, ...)
}
\arguments{
\item{type}{data}
}
\description{
Measure Matrix
}
\author{
M.N. Legasa
}
