% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal2UniformCorrelation.R
\name{normal2UniformCorrelation}
\alias{normal2UniformCorrelation}
\title{Find Uniform Correlation from Normal Correlation}
\usage{
normal2UniformCorrelation(r, p01e1, p11e1, p01e2, p11e2,
  accuracy = 1e+07)
}
\description{
Find Uniform Correlation from Normal Correlation
}
\author{
M.N. Legasa
}
