% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelHandler.R
\name{parallelHandler}
\alias{parallelHandler}
\title{Parallelization Cluster Starter}
\usage{
parallelHandler(parallelize = FALSE, type = "PSOCK", n.cores = NULL,
  cl = NULL, PSOCK.Exports.list = list())
}
\arguments{
\item{type}{Cluster type, either FORK or PSOCK.}

\item{n.cores}{Number of cores.}

\item{PSOCK.Exports.list}{For PSOCK type cluster, named list of
functions to be exported. It requires each function and variable name name to be 
 specified in the list names. Use \Code{namedList()} if necessary.}
}
\description{
Parallelization Cluster Starter
}
\author{
M.N. Legasa
}
