% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toOperableMatrix.R
\name{toOperableMatrix}
\alias{toOperableMatrix}
\title{Convert Valid Object to Operable Matrix}
\usage{
toOperableMatrix(data)
}
\description{
Convert Valid Object to Operable Matrix
}
\author{
M.N. Legasa
}
