% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilksGenerateSeries.R
\name{wilksGenerateSeries}
\alias{wilksGenerateSeries}
\title{Generate Series from Wilks Weather Generator}
\usage{
wilksGenerateSeries(wilks, n, initial = NULL, initial.date = NULL,
  tol = 10^(-2))
}
\description{
Generate Series from Wilks Weather Generator
}
\author{
M.N. Legasa
}
