% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilksTrain.R
\name{wilksTrain}
\alias{wilksTrain}
\title{Train Complete Wilks Model}
\usage{
wilksTrain(data, accuracy = 1e+07, max.error = 0.001, max.iter = 20,
  parallelize = TRUE, n.cores = NULL)
}
\description{
Train Complete Wilks Model
}
\author{
M.N. Legasa
}
